WIP: get more schemas in validation loop
Now that the basic mechanics are working introduce more schemas into the
validation chain.

These changes were necessary in order to allow the mechanics of
loading and validating the schemas to work correctly.

Specifically, until we can refactor the schemas to be efficient about
re-using data types, there must be no overlapping id's in different
definitions or the validator will choke loading the schemas with an
error about schema already defined with a different definition.

To just allow us to see the mechanics of validation working, I
adjusted all ids to be different, basically.