Fix tool to run with any size file
This causes the process to wait for all the json to parse before
attempting to validate it; before it would attempt to incrementally
parse and validate json.  This would cause errors when attempting to
validate large chunks of valid json, which needs to be parsed as one
block.

Fixes issue reported by @jebeck, thanks!