require that value be a number, kill spurious prop
Each record had a custom property that duplicated value, but using the
name (the type) of the record as the key name for the duplicate value.
In the short term this was used to help develop the data model and get
it used in tests throughout the parsers and experiment with numeric vs
string types.  Now that we know we want numbers everywhere, we can
safely assume that there are no objects which have string "values".

Now the data model can use the $ref feature to abstract a common base
record.  All records have a type, deviceTime, and value.  Value is
always a number.